requirements.txt
torch==1.8.1
librosa==0.8.0
numpy==1.20.2
scikit-learn==0.24.1
scipy==1.6.2
matplotlib==3.4.1
webrtcvad==2.0.10
Flask==1.1.2
Werkzeug==1.0.1
Flask-JWT==0.3.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
python-dotenv==0.17.0